Fix #4 by selecting a subset of buffers
In case not all buffers can be displayed at once
select a subset of buffers and display them.
The subset is not always optimal because we
can only approximate the space available and
space consumed by the buffers.

In any case the result is better than before because
the currently selected buffer is always visible.