Automatically show tabline based on buffer count
A new configuration value - `g:lightline#bufferline#min_buffer_count` -
controls this behavior. When set to a positive value, it represents the
minimum number of buffers needed to automatically show the tabline; the
tabline is hidden when the buffer count is less. When set to 0 (the
default), the automatic tabline behavior is disabled.

This change requires the introduction of a "plugin" script in order to
trigger the initialization of a new 'lightline_bufferline' autocommand
group. This unfortunately partially defeats the purpose of the autoload
script, but I wanted to introduce the minimum number of changes here.

The code could be reorganized to be more autoload-friendly by defining
only functions in the autoload script and moving all configuration and
mappings to the plugin script, but that can be done in future changes.