Delay the event handler to fix bufferline event updates
When unloading a buffer in a vertical split the unloaded buffer is
still listed as loaded by vim. As a workaround the event
handler is delayed a few milliseconds using a timer.
A nice side effect is that the event handler is not called multiple
times in a row because it will only be called once every 10ms.

Fixes #82.